% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgcidr-package.R
\docType{package}
\name{pgcidr}
\alias{pgcidr}
\alias{pgcidr-package}
\title{Perform IP Address in CIDR Lookups with PostgreSQL}
\description{
PostgreSQL has built-in support for IP address and CIDR types
but the ip4r \url{https://github.com/RhodiumToad/ip4r} extension is much faster.
Tools are provided to create CIDR lookup and IP address source tables and
perform IP address in CIDR queries. Documentation on how to setup macOS
with PostgreSQL and ip4r is also provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/pgcidr}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/pgcidr/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
